+++ src/ts/utils/datastructures/CircularIntervalSet.ts
@@ -1,13 +1,14 @@
 /**
- * A circular interval set.
+ * @classdesc A circular interval set.
  * 
  * @author   Ikaros Kappler
  * @date     2020-10-02
  * @modified 2020-10-18 Ported to Typescript from vanilla JS.
  * @version  1.0.0
+ * @name CircularIntervalSet
  **/
 
-class CircularIntervalSet {
+export class CircularIntervalSet {
 
     /**
      * @member {number} start